Fail fast on snapshot / release version mismatch upon cloning
without this change we will know that Spring Cloud Release has wrong setup only when we do POM updates
with this change, upon cloning the Spring Cloud Release project we will immediately check that there are snapshot versions for a release of a non-snapshot version of a project. If that's the case then we throw an exception.

fixes #42